More tag rework, trying other way to sort the xml away

+++ service/internal/sourcer/abletonv5/device_chain_struct.go
@@ -4,7 +4,7 @@ type DeviceChainDocument struct {
 	HasBase
 	HasFileReference
 
-	DeviceCount uint64 `json:"device_count"`
+	DeviceCount int `json:"device_count"`
 }
 
 func NewDeviceChainDocument() *DeviceChainDocument {
